cleanup: consolidate random utilities into one file
