+++ home/templates/home/home.html
@@ -4,7 +4,7 @@
 
 {% block content %}
 
-    <h1>Elizabeth Schley's Personal Study and Fun Site</h1>
+    <h1>Study and Fun Site</h1>
     <div class="clearfix">
         <a href="#">
             <img class="default float-right"
@@ -12,7 +12,7 @@
                     alt='Fort Tryon Park, NYC'>
         </a>
         <p>This site is all about learning and sharing information. Although I have many ideas for the future, for now it is a portfolio and study aid.</p>
-        <p>When I am an employed programmer, work always comes first.  Yet, I always imagine applications that I would create, if my priorities were different.  This site will eventually give me the platform to do that.  But in the here and now, my first goal is to create a site, allowing me to learn whatever I want, whenever I want, and to share the information with anyone interested. </p>
+        <p>When I am an employed programmer, work always comes first.  Yet, I always imagine applications to create.  This site will eventually give me the platform to do that.  But in the here and now, my first goal is to create a site, helping me learn whatever I want, whenever I want, and to share the information with anyone interested. </p>
         <p>Note about <a href="#" data-slug="home-page-images" class="one_para_modal modal_popup_link">home page images</a></p>
      </div>
      <div class="row">
@@ -32,8 +32,8 @@
             <li>No typing https to get https and no www ever (it won't work)</li>
             <li>No extra buttons when a refresh or back button will work just as well.</li>
             <li>No database updates using a form as input</li>
-            <li>Only a few static pages</li>
-            <li>Data driven display: most pages will pull the data from the database, allowing the same information to be displayed on a page in many ways, based on the category, group and/or paragraph attributes. For example, the same paragraph can have a page or pop-up of its own, or can be part of an ordered group, such as an answer to the a flashcard question</li>
+            <li>Mostly dynamic content</li>
+            <li>Data driven display: most pages will pull the data from the database, allowing the same information to be displayed on a page in many ways, based on the category, group and/or paragraph attributes. For example, the same paragraph can have a page or pop-up of its own, or can be part of an ordered group, such as an answer to a flashcard question or part of step by step instructions</li>
             <li>Easy and convenient to upload data and static files</li>
             <li>Excellent automated tests (the unmet requirement)</li>
         </ol>
